change 'submission' to 'script'
This is a big change as it modifies the DB, renames files, classnames
and generally touches a lot of things. Everywhere 'submission' as a
datatype or path is mentioned has been changed to be 'script' instead.

Also from a user's perspective, the language used is more about job
settings than submissions.