determine start and end frames in ruby
This fixes at least 1 bug where start and end frame calculations in
bash were incorrect if many nodes (more than 1/2 of frames) were used
and most frames would be rendered in the last task. This also moves
creating thumbnails to ruby so that we only create thumbnails for
for the frames that were generated in that task. Optimizing thumbnail
creating but also ensuring there's no race condition on converting
incomplete frames and throwing misleading errors.