bugfix for end_frames including start frame
There was a bug where end_frames where not exactly counted right
and this was particularly bad if the task_size was 2. This fix
now includes the start_frame in the task_size when determining end_frame.
I.e., a start_frame of 5 + task_size of 5 returned and end_frame of 10,
but that actually should be 9 becuase frame 5 is included in the task.